Fixed test [skip hazardlib]
